+++ clients/notification-kafka/src/main/java/org/eclipse/hono/client/notification/kafka/NotificationTopicHelper.java
@@ -25,6 +25,11 @@ import org.eclipse.hono.notification.deviceregistry.TenantChangeNotification;
  */
 public final class NotificationTopicHelper {
 
+    private static final String TENANT_CHANGE_TOPIC = getTopicNameForAddress(TenantChangeNotification.ADDRESS);
+    private static final String DEVICE_CHANGE_TOPIC = getTopicNameForAddress(DeviceChangeNotification.ADDRESS);
+    private static final String CREDENTIALS_CHANGE_TOPIC = getTopicNameForAddress(CredentialsChangeNotification.ADDRESS);
+    private static final String ALL_DEVICES_OF_TENANT_CHANGE_TOPIC = getTopicNameForAddress(AllDevicesOfTenantDeletedNotification.ADDRESS);
+
     private NotificationTopicHelper() {
         // prevent instantiation
     }
@@ -38,20 +43,19 @@ public final class NotificationTopicHelper {
      * @throws IllegalArgumentException If the given type is not a known subclass of {@link AbstractNotification}.
      */
     public static <T extends AbstractNotification> String getTopicName(final Class<T> notificationType) {
-        final String address;
+        final String topic;
         if (TenantChangeNotification.class.equals(notificationType)) {
-            address = TenantChangeNotification.ADDRESS;
+            topic = TENANT_CHANGE_TOPIC;
         } else if (DeviceChangeNotification.class.equals(notificationType)) {
-            address = DeviceChangeNotification.ADDRESS;
+            topic = DEVICE_CHANGE_TOPIC;
         } else if (CredentialsChangeNotification.class.equals(notificationType)) {
-            address = CredentialsChangeNotification.ADDRESS;
+            topic = CREDENTIALS_CHANGE_TOPIC;
         } else if (AllDevicesOfTenantDeletedNotification.class.equals(notificationType)) {
-            address = AllDevicesOfTenantDeletedNotification.ADDRESS;
+            topic = ALL_DEVICES_OF_TENANT_CHANGE_TOPIC;
         } else {
             throw new IllegalArgumentException("Unknown notification type " + notificationType.getName());
         }
-
-        return getTopicNameForAddress(address);
+        return topic;
     }
 
     private static String getTopicNameForAddress(final String address) {